refactor(customizer): extract shared popover chrome from typography
Move openPopover/closePopover/positionPopover/bindDismiss/unbindDismiss
and the one-popover-at-a-time manager out of typography-control.js into
popover-chrome.js as an attachPopoverChrome(runtime, opts) mixin — the
code is verbatim; only the anchor element (opts.anchor) and an optional
onClose hook are parameterized. The module-scoped activePopover ref is
now shared by every runtime the factory touches, so popovers stay
exclusive ACROSS control types once styling adopts the chrome.
