+++ src/backend/customizer/js/typography-control.js
@@ -15,6 +15,8 @@
  * Other refs (`_`, `Customify_Control_Args`, `_wpCustomizeSettings`)
  * are genuine browser globals and read directly.
  */
+import { attachPopoverChrome } from "./popover-chrome";
+
 export function setupTypographyControl(deps) {
 	var $ = deps.$;
 	var $document = deps.$document;
@@ -137,9 +139,6 @@ export function setupTypographyControl(deps) {
 		$(".customify-trigger--family", $trigger).text(family);
 		$(".customify-trigger--meta", $trigger).text(meta.join(" / "));
 	}
-
-	// Only one typography popover may be open at a time.
-	var activePopover = null;
 
 	var FontSelector = {
 		fonts: null,
@@ -591,143 +590,6 @@ export function setupTypographyControl(deps) {
 			}
 		},
 
-		// ── Popover chrome ─────────────────────────────────────────────
-		// The settings panel floats over the controls below (absolute
-		// within the control li) instead of the legacy inline accordion.
-		// Open/close is class-driven (`is-open`) so the CSS transition in
-		// _control.scss animates opacity/transform; jQuery slide* would
-		// jump because it animates height inline.
-		openPopover: function () {
-			var that = this;
-			if (activePopover && activePopover !== that) {
-				activePopover.closePopover();
-			}
-			activePopover = that;
-
-			that.$el.attr("data-opening", "opening");
-			that.$el.addClass("modal--opening");
-			$(".action--reset", that.$el).show();
-			that.positionPopover();
-
-			// A freshly-appended panel needs its initial (hidden) styles
-			// committed before the class flips, or the transition jumps
-			// straight to the final state. Force a synchronous reflow
-			// instead of requestAnimationFrame: rAF never fires while the
-			// tab is hidden (background window / tab switch), so the
-			// class would land arbitrarily late — even after a
-			// closePopover() — resurrecting a dismissed popover.
-			if (that.container[0]) {
-				void that.container[0].offsetWidth;
-			}
-			that.container.addClass("is-open");
-			that.bindDismiss();
-		},
-
-		closePopover: function () {
-			var that = this;
-			var $el = that.$el;
-			if (activePopover === that) {
-				activePopover = null;
-			}
-			that.unbindDismiss();
-			$el.attr("data-opening", "");
-			$el.removeClass("modal--opening");
-			$(".action--reset", $el).hide();
-			if (that.container) {
-				that.container.removeClass("is-open");
-			}
-		},
-
-		// Anchor the popover right under the trigger (or above it when
-		// the viewport space below is too small). `top` is relative to
-		// the control li — the positioning context set by the
-		// `customify-modal--inside` class.
-		positionPopover: function () {
-			var that = this;
-			var $trigger = $(".customify-typo-trigger", that.$el);
-			if (
-				!$trigger.length ||
-				!that.container ||
-				!that.container.length
-			) {
-				return;
-			}
-			var GAP = 8;
-			var liRect = that.$el[0].getBoundingClientRect();
-			var tRect = $trigger[0].getBoundingClientRect();
-			var popH = that.container.outerHeight();
-			var spaceBelow = window.innerHeight - tRect.bottom;
-			var top;
-			if (spaceBelow < popH + 24 && tRect.top - popH - 24 > 0) {
-				top = tRect.top - liRect.top - popH - GAP;
-				that.container.addClass("is-above");
-			} else {
-				top = tRect.bottom - liRect.top + GAP;
-				that.container.removeClass("is-above");
-			}
-			that.container.css("top", Math.round(top) + "px");
-		},
-
-		// Close on outside click / ESC while open. Native capture-phase
-		// listeners: outside-click must win over handlers that
-		// stopPropagation, and ESC must fire before the Customizer's own
-		// document-level ESC (which would also collapse the section).
-		bindDismiss: function () {
-			var that = this;
-			that.unbindDismiss();
-			that._onOutside = function (e) {
-				if ($(e.target).closest(that.$el[0]).length) {
-					return;
-				}
-				// Select2 renders its dropdown at <body> level — clicks
-				// there belong to the popover even though they land
-				// outside the control li.
-				if (
-					$(e.target).closest(
-						".select2-container, .select2-dropdown"
-					).length
-				) {
-					return;
-				}
-				that.closePopover();
-			};
-			that._onEsc = function (e) {
-				if ("Escape" === e.key || 27 === e.keyCode) {
-					e.stopPropagation();
-					that.closePopover();
-				}
-			};
-			// Clicks inside the preview iframe never reach this document's
-			// mousedown listener — but focusing the iframe blurs the top
-			// window, so treat window blur as an outside click.
-			that._onBlur = function () {
-				that.closePopover();
-			};
-			document.addEventListener("mousedown", that._onOutside, true);
-			document.addEventListener("keydown", that._onEsc, true);
-			window.addEventListener("blur", that._onBlur);
-		},
-
-		unbindDismiss: function () {
-			var that = this;
-			if (that._onOutside) {
-				document.removeEventListener(
-					"mousedown",
-					that._onOutside,
-					true
-				);
-				that._onOutside = null;
-			}
-			if (that._onEsc) {
-				document.removeEventListener("keydown", that._onEsc, true);
-				that._onEsc = null;
-			}
-			if (that._onBlur) {
-				window.removeEventListener("blur", that._onBlur);
-				that._onBlur = null;
-			}
-		},
-
 		reset: function () {
 			//this.$el = $el;
 			var that = this;
@@ -808,6 +670,17 @@ export function setupTypographyControl(deps) {
 			this.load();
 		}
 	};
+
+	// Popover lifecycle (open/close/position/dismiss) comes from the
+	// shared chrome — one popover at a time across typography AND styling
+	// controls. Attached to the base object so the per-control clones
+	// made in intTypos() inherit the methods.
+	attachPopoverChrome(FontSelector, {
+		$: $,
+		anchor: function (that) {
+			return $(".customify-typo-trigger", that.$el);
+		}
+	});
 
 	var intTypoControls = {};
 	var intTypos = function () {
